Toolbox/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
imageio==2.8.0
numpy==1.21.2
Pillow==8.3.1
PyYAML==5.4.1
lxml==4.6.3
svgpathtools==1.4.1
codepoints==1.0
